Add select all repositories action
