+++ app/components/owner/migrate/webhooks.js
@@ -1,26 +1,5 @@
 import Component from '@ember/component';
+import SelectableRepositoriesList from 'travis/mixins/components/selectable-repositories-list';
 
-export default Component.extend({
-
-  selectedRepositories: null,
-
-  init() {
-    this.set('selectedRepositories', []);
-    return this._super(...arguments);
-  },
-
-  actions: {
-
-    toggleRepository(repo) {
-      const { selectedRepositories } = this;
-      const isSelected = selectedRepositories.includes(repo);
-      if (isSelected) {
-        selectedRepositories.removeObject(repo);
-      } else {
-        selectedRepositories.addObject(repo);
-      }
-    }
-
-  }
-
+export default Component.extend(SelectableRepositoriesList, {
 });
